refactor: move collector data to upper scope
